requirements.txt
google-cloud-automl
google-cloud-storage